Refactor application to handle all data loading
This is a significant refactor that changes the role of slingshot. The
application will now handle loading and publishing all layers. Rather
than relying on GeoServer to load the vector data into PostGIS, which
was always slow and unreliable, the layers are now loaded directly into
PostGIS and later registered in GeoSever.