Adds support for processing raster data
This adds a `normalize` function that accepts a GeoTIFF file as input
and creates a new GeoTIFF which has been optimized for serving through
GeoServer. The following steps are taken:

* Paletted images are expanded to RGB.
* Images are internally tiled using a block size of 512.
* Overview images are added to the TIFF when larger than block size.

Support is provided through the `rasterio` package which itself depends
on GDAL. You'll need to make sure the header files for GDAL are
accessible in order to install it.