Add GeoTIFFs to workflow
This adds support for GeoTIFFs. There's still an issue of where to put
the rasters for downloads. For now the approach is just to copy the TIFF
from the NFS mount to a directory served by Apache. A better solution
involving, most likely, S3 should be considered for the future.