Correctly handle layers with multiple XML files
Some data layers can legitimately contain more than one XML file. We
need to be able to differentiate between them in order to locate the
FGDC metadata. This provides some basic support to do this. A more
robust implementation should incrementally parse the beginning of an XML
file to find the root element.