Refactors the record object
The old record object was difficult to maintain and easy to break. This
has been replaced with an object based off of the third party `attrs`
module. This new record requires slightly more effort to use, but it's
much more difficult to misuse or make mistakes while modifying.

The implementation itself provides an almost entirely similar interface
for creation. The main differences are:

* Immutability; it is no longer possible to change a field once the
record has been created. (It's still possible to modify mutable fields
like dictionaries or lists.)
* Datetime fields must be correctly formatted before creation.
* The `solr_geom` field should be assembled before creating the record.