Add support for PG schemas
This adds support for putting the data layers into a separate PG schema.
It's a bit of a hack in its current state as the schema is hard coded as
``geodata``. This will need to be fixed to make it configurable. The
tests will also need to be updated to reflect this change.

The addition of schema support is necessary because both the spatial
data and the Rails app currently share the same table space. The better
approach will be to pull them out into separate spaces.