Skip already published layers
This assumes that the existence of an FGDC metadata file in the metadata
directory means a layer is already published. Under normal operation
this assumption should be true. The alternative would be to also check
whether the layer exists in GeoServer and Solr, but this would result in
thousands of HTTP requests.