Switch from file system to S3
This is a significant rewrite in preparation for moving to AWS. The
biggest change is to move from dealing with layers stored on the file
system to layers stored in S3. GeoTIFFs will be stored and accessed by
GeoServer directly from S3, so the process for publishing GeoTIFFs has
changed.

The minimum required Python version is now 3.7. This is needed to take
advantage of seekable zipfiles.