Add docker build
This adds a Dockerfile for building the image. The wheel packaging step
has been moved into a docker stage for more consistency in the
build/deploy pipeline.

The git versioning for the app has been removed. It was always kind of
fragile and is more work to do in the container than it's probably
worth. git versions will be applied to the container tag during the
continuous integration process, so versioning will be managed at the
container level.