Support publishing multiple layers at once
This changes the publish command to support publishing multiple layers
either by specifying the uploaded layer names as arguments or by using
the new `--publish-all` switch. If using the latter, this will scan the
upload bucket and publish all layers it finds. If a layer has already
been published it will be skipped if the published layer is newer than
the uploaded layer.

Rudimentary information about a layer's state is managed in a DynamoDB
table.