Add support for components of database URL
This adds options to the publish command for configuring the database by
specifying each of the URL components individually. The old behavior of
using the single URL string is still supported; you can now use one or
the other.

It's way easier in Terraform to configure the Fargate task using the
separate URL components.