Make FGDC publicly available in S3
The FGDC file for each layer should be made public, regardless of
whether or not the layer itself is restricted. This adds a step during
the publishing process to change the ACL to public read on the FGDC file
as well as adding the necessary link to the file in the GeoBlacklight
record.