Wait for ECS to flush logs
It turns out ECS can frequently discard the logs before fully flushing
them to Cloudwatch. This was supposedly fixed recently but my experience
suggests otherwise. This adds a 10 second pause after slingshot has
exited before fully exiting the container. It seems to be enough to
ensure that all the slingshot logs are written to Cloudwatch.