Use io.BufferedReader for Postgres COPY
This fixes a few problems with the S3IO wrapper. These changes make it
possible to use the `io.BufferedReader` for the different files in the
Shapefile when loading into Postgres. Memory usage should now be fully
capped and largely controlled by the number of worker threads and size
of the `S3_BUFFER_SIZE` constant.