Switch MARC loading to use MARC21 from S3
The old process that was used to load MARC records relied on a custom
script to generate some MARCXML that was then SCPed to the server GeoWeb
was running on. With the move to AWS, instead of trying to modify the
old process, this just uses the MARC that's uploaded to S3 for the Mario
project. The actual loading process will be implemented in Airflow.