Batch the Solr add calls for MARC records
This batches the MARC records into collections of documents instead of
issuing a Solr call per record.

It also switches to using a more forgiving MARC parser than what comes
with pymarc. A lot of our MARC data is kind of broken and pymarc does
not handle it well at all.