Fix bug in MARC parser, update Makefile
Why these changes are being introduced:

The record publication date field was processed into JSON as an array
instead of a string, which resulted in very strange dates in the
Geoblacklight view. The dates also include some odd characters that are
correct for MARC cataloging but that we don't want to display.

How this addresses that need:
* Changs the record date field (dct_temporal_sm) to a string instead of
  an array
* Strips brackets and other cataloging characters
  year string before ingesting
* Also updates the Makefile to use new AWS CLI login command

Side effects of this change:
None, but Slingshot will need to be re-run after merge with the latest
full export from Alma.